+++ src/QueryType/ManagedResources/Query/AbstractCommand.php
@@ -29,11 +29,4 @@ abstract class AbstractCommand extends Configurable
      * @return string
      */
     abstract public function getRequestMethod(): string;
-
-    /**
-     * Returns the data to be sent to Solr.
-     *
-     * @return string
-     */
-    abstract public function getRawData(): string;
 }