+++ src/QueryType/ManagedResources/Query/AbstractQuery.php
@@ -57,11 +57,18 @@ abstract class AbstractQuery extends BaseQuery implements Status4xxNoExceptionIn
     protected $commandTypes;
 
     /**
-     * Name of the managed resource.
+     * Name of the managed resource to query.
      *
      * @var string
      */
     protected $name = '';
+
+    /**
+     * Name of the child resource to query.
+     *
+     * @var string|null
+     */
+    protected $term = null;
 
     /**
      * Default result class if no command is set.
@@ -121,6 +128,42 @@ abstract class AbstractQuery extends BaseQuery implements Status4xxNoExceptionIn
     public function setName(string $name): self
     {
         $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * Get the name of the child resource to query.
+     *
+     * @return string|null
+     */
+    public function getTerm(): ?string
+    {
+        return $this->term;
+    }
+
+    /**
+     * Set the name of the child resource to query.
+     *
+     * @param string $term
+     *
+     * @return self
+     */
+    public function setTerm(string $term): self
+    {
+        $this->term = $term;
+
+        return $this;
+    }
+
+    /**
+     * Remove the name of the child resource. This reverts to querying the entire managed resource.
+     *
+     * @return self
+     */
+    public function removeTerm(): self
+    {
+        $this->term = null;
 
         return $this;
     }