+++ src/QueryType/ManagedResources/Query/Synonyms.php
@@ -20,7 +20,8 @@ use Solarium\QueryType\ManagedResources\Query\Command\Synonyms\Create;
 use Solarium\QueryType\ManagedResources\Query\Synonyms\InitArgs;
 use Solarium\QueryType\ManagedResources\ResponseParser\Command as CommandResponseParser;
 use Solarium\QueryType\ManagedResources\ResponseParser\Exists as ExistsResponseParser;
-use Solarium\QueryType\ManagedResources\ResponseParser\Synonyms as ResponseParser;
+use Solarium\QueryType\ManagedResources\ResponseParser\Synonym as SynonymResponseParser;
+use Solarium\QueryType\ManagedResources\ResponseParser\Synonyms as SynonymsResponseParser;
 use Solarium\QueryType\ManagedResources\Result\Synonyms\SynonymMappings;
 
 /**
@@ -73,12 +74,17 @@ class Synonyms extends AbstractQuery
     /**
      * Get the response parser class for this query.
      *
-     * @return \Solarium\QueryType\ManagedResources\ResponseParser\Synonyms|\Solarium\QueryType\ManagedResources\ResponseParser\Command
+     * @return \Solarium\Core\Query\ResponseParserInterface
      */
     public function getResponseParser(): ResponseParserInterface
     {
         if (null === $this->command) {
-            $parser = new ResponseParser();
+            if (null === $this->term) {
+                $parser = new SynonymsResponseParser();
+            }
+            else {
+                $parser = new SynonymResponseParser();
+            }
         } elseif (self::COMMAND_EXISTS === $this->command->getType()) {
             $parser = new ExistsResponseParser();
         } else {