+++ src/QueryType/ManagedResources/RequestBuilder/Resource.php
@@ -42,8 +42,12 @@ class Resource extends AbstractRequestBuilder
             $request->addHeader('Content-Type: application/json; charset=utf-8');
             $this->buildCommand($request, $query->getCommand());
         } else {
-            // Lists all items.
+            // Lists one or all items.
             $request->setMethod(Request::METHOD_GET);
+
+            if (null !== $term = $query->getTerm()) {
+                $request->setHandler($request->getHandler().'/'.$term);
+            }
         }
 
         return $request;
@@ -63,13 +67,22 @@ class Resource extends AbstractRequestBuilder
 
         switch ($command->getType()) {
             case BaseQuery::COMMAND_ADD:
-                $request->setRawData($command->getRawData());
+                if (null === $rawData = $command->getRawData()) {
+                    throw new RuntimeException('Missing data for ADD command.');
+                }
+                $request->setRawData($rawData);
                 break;
             case BaseQuery::COMMAND_CONFIG:
-                $request->setRawData($command->getRawData());
+                if (null === $rawData = $command->getRawData()) {
+                    throw new RuntimeException('Missing initArgs for CONFIG command.');
+                }
+                $request->setRawData($rawData);
                 break;
             case BaseQuery::COMMAND_CREATE:
-                $request->setRawData($command->getRawData());
+                if (null === $rawData = $command->getRawData()) {
+                    throw new RuntimeException('Missing class for CREATE command.');
+                }
+                $request->setRawData($rawData);
                 break;
             case BaseQuery::COMMAND_DELETE:
                 if (null === $term = $command->getTerm()) {