+++ src/QueryType/ManagedResources/ResponseParser/Stopword.php
@@ -14,9 +14,9 @@ use Solarium\Core\Query\ResponseParserInterface;
 use Solarium\Core\Query\Result\ResultInterface;
 
 /**
- * Stopwords.
+ * Stopword.
  */
-class Stopwords extends ResponseParserAbstract implements ResponseParserInterface
+class Stopword extends ResponseParserAbstract implements ResponseParserInterface
 {
     /**
      * Parse response data.
@@ -27,29 +27,21 @@ class Stopwords extends ResponseParserAbstract implements ResponseParserInterfac
      */
     public function parse(ResultInterface $result): array
     {
-        $wordSet = null;
         $data = [];
         $parsed = ['items' => []];
         $parsed = $this->parseStatus($parsed, $result);
 
         if ($parsed['wasSuccessful']) {
             $data = $result->getData();
-            if (isset($data['wordSet'])) {
-                $wordSet = $data['wordSet'];
-            }
-        }
+            $items = [];
 
-        if (null !== $wordSet && !empty($wordSet)) {
-            $parsed['items'] = $wordSet['managedList'];
-            $parsed['initializedOn'] = $wordSet['initializedOn'];
-
-            if (isset($wordSet['initArgs']['ignoreCase'])) {
-                $parsed['ignoreCase'] = $wordSet['initArgs']['ignoreCase'];
+            foreach ($data as $term => $stopword) {
+                if ('responseHeader' !== $term) {
+                  $items[] = $stopword;
+                }
             }
 
-            if (isset($wordSet['updatedSinceInit'])) {
-                $parsed['updatedSinceInit'] = $wordSet['updatedSinceInit'];
-            }
+            $parsed['items'] = $items;
         }
 
         $parsed = $this->addHeaderInfo($data, $parsed);