+++ tests/QueryType/ManagedResources/Query/StopwordsTest.php
@@ -16,7 +16,8 @@ use Solarium\QueryType\ManagedResources\Query\Stopwords\InitArgs;
 use Solarium\QueryType\ManagedResources\RequestBuilder\Resource as RequestBuilder;
 use Solarium\QueryType\ManagedResources\ResponseParser\Command as CommandResponseParser;
 use Solarium\QueryType\ManagedResources\ResponseParser\Exists as ExistsResponseParser;
-use Solarium\QueryType\ManagedResources\ResponseParser\Stopwords as ResponseParser;
+use Solarium\QueryType\ManagedResources\ResponseParser\Stopword as StopwordResponseParser;
+use Solarium\QueryType\ManagedResources\ResponseParser\Stopwords as StopwordsResponseParser;
 
 class StopwordsTest extends TestCase
 {
@@ -39,7 +40,13 @@ class StopwordsTest extends TestCase
 
     public function testGetResponseParser()
     {
-        $this->assertInstanceOf(ResponseParser::class, $this->query->getResponseParser());
+        $this->assertInstanceOf(StopwordsResponseParser::class, $this->query->getResponseParser());
+    }
+
+    public function testGetResponseParserWithTerm()
+    {
+        $this->query->setTerm('test');
+        $this->assertInstanceOf(StopwordResponseParser::class, $this->query->getResponseParser());
     }
 
     public function testGetResponseParserWithCommand()
@@ -47,8 +54,6 @@ class StopwordsTest extends TestCase
         $command = $this->query->createCommand(Stopwords::COMMAND_ADD);
         $this->query->setCommand($command);
         $this->assertInstanceOf(CommandResponseParser::class, $this->query->getResponseParser());
-        $this->query->removeCommand();
-        $this->assertInstanceOf(ResponseParser::class, $this->query->getResponseParser());
     }
 
     public function testGetResponseParserWithExistsCommand()
@@ -56,8 +61,33 @@ class StopwordsTest extends TestCase
         $command = $this->query->createCommand(Stopwords::COMMAND_EXISTS);
         $this->query->setCommand($command);
         $this->assertInstanceOf(ExistsResponseParser::class, $this->query->getResponseParser());
+    }
+
+    public function testGetResponseParserAfterRemovingCommand()
+    {
+        $command = $this->query->createCommand(Stopwords::COMMAND_ADD);
+        $this->query->setCommand($command);
         $this->query->removeCommand();
-        $this->assertInstanceOf(ResponseParser::class, $this->query->getResponseParser());
+        $this->assertInstanceOf(StopwordsResponseParser::class, $this->query->getResponseParser());
+    }
+
+    public function testSetAndGetName()
+    {
+        $this->query->setName('test');
+        $this->assertSame('test', $this->query->getName());
+    }
+
+    public function testSetAndGetTerm()
+    {
+        $this->query->setTerm('test');
+        $this->assertSame('test', $this->query->getTerm());
+    }
+
+    public function testRemoveTerm()
+    {
+        $this->query->setTerm('test');
+        $this->query->removeTerm();
+        $this->assertNull($this->query->getTerm());
     }
 
     public function testCommand()
@@ -120,30 +150,8 @@ class StopwordsTest extends TestCase
 
     public function testCreateInitArgsWithArgs()
     {
-        $args = ['ignoreCase' => true];
-        $initArgs = $this->query->createInitArgs($args);
+        $initArgs = $this->query->createInitArgs(['ignoreCase' => true]);
         $this->assertInstanceOf(InitArgs::class, $initArgs);
-        $this->assertEquals($args, $initArgs->getInitArgs());
-    }
-
-    public function testInitArgs()
-    {
-        $initArgs = new InitArgs();
-        $this->assertEquals([], $initArgs->getInitArgs());
-    }
-
-    public function testInitArgsIgnoreCase()
-    {
-        $initArgs = new InitArgs();
-        $initArgs->setIgnoreCase(true);
-        $this->assertTrue($initArgs->getIgnoreCase());
-    }
-
-    public function testInitArgsSet()
-    {
-        $args = ['ignoreCase' => true];
-        $initArgs = new InitArgs();
-        $initArgs->setInitArgs($args);
-        $this->assertEquals($args, $initArgs->getInitArgs());
+        $this->assertEquals(['ignoreCase' => true], $initArgs->getInitArgs());
     }
 }