+++ tests/QueryType/ManagedResources/RequestBuilder/StopwordsTest.php
@@ -31,14 +31,13 @@ class StopwordsTest extends TestCase
     public function setUp(): void
     {
         $this->query = new StopwordsQuery();
+        $this->query->setName('dutch');
         $this->builder = new StopwordsRequestBuilder();
     }
 
     public function testBuild()
     {
         $handler = 'schema/analysis/stopwords/dutch';
-
-        $this->query->setName('dutch');
         $request = $this->builder->build($this->query);
 
         $this->assertEquals(
@@ -55,6 +54,7 @@ class StopwordsTest extends TestCase
 
     public function testNoName()
     {
+        $this->query->setName('');
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage('Name of the resource is not set in the query.');
         $request = $this->builder->build($this->query);
@@ -63,9 +63,8 @@ class StopwordsTest extends TestCase
     public function testUnsupportedCommand()
     {
         $command = new UnsupportedStopwordsCommand();
-        $this->query->setName('dutch');
-        $this->query->setCommand($command);
 
+        $this->query->setCommand($command);
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage('Unsupported command type: unsupportedtype');
         $request = $this->builder->build($this->query);
@@ -73,22 +72,26 @@ class StopwordsTest extends TestCase
 
     public function testQuery()
     {
-        $this->query->setName('dutch');
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_GET, $request->getMethod());
         $this->assertSame('schema/analysis/stopwords/dutch', $request->getHandler());
         $this->assertNull($request->getRawData());
     }
 
+    public function testQueryWithTerm()
+    {
+        $this->query->setTerm('de');
+        $request = $this->builder->build($this->query);
+        $this->assertSame(Request::METHOD_GET, $request->getMethod());
+        $this->assertSame('schema/analysis/stopwords/dutch/de', $request->getHandler());
+        $this->assertNull($request->getRawData());
+    }
+
     public function testAdd()
     {
-        $stopwords = ['de'];
         $command = new AddCommand();
-        $command->setStopwords($stopwords);
-        $this->assertSame($stopwords, $command->getStopwords());
-        $this->assertSame('["de"]', $command->getRawData());
+        $command->setStopwords(['de']);
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_PUT, $request->getMethod());
@@ -96,21 +99,24 @@ class StopwordsTest extends TestCase
         $this->assertSame('["de"]', $request->getRawData());
     }
 
+    public function testAddWithoutStopwords()
+    {
+        $command = new AddCommand();
+
+        $this->query->setCommand($command);
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('Missing data for ADD command.');
+        $request = $this->builder->build($this->query);
+    }
+
     public function testConfig()
     {
         $initArgs = new InitArgs();
-        $command = new ConfigCommand();
-
-        $command->setInitArgs($initArgs);
-        $this->assertSame($initArgs, $command->getInitArgs());
-        $this->assertSame('', $command->getRawData());
-
         $initArgs->setInitArgs(['ignoreCase' => true]);
-        $command->setInitArgs($initArgs);
-        $this->assertSame($initArgs, $command->getInitArgs());
-        $this->assertSame('{"initArgs":{"ignoreCase":true}}', $command->getRawData());
 
-        $this->query->setName('dutch');
+        $command = new ConfigCommand();
+        $command->setInitArgs($initArgs);
+
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_PUT, $request->getMethod());
@@ -118,12 +124,20 @@ class StopwordsTest extends TestCase
         $this->assertSame('{"initArgs":{"ignoreCase":true}}', $request->getRawData());
     }
 
+    public function testConfigWithoutInitArgs()
+    {
+        $command = new ConfigCommand();
+
+        $this->query->setCommand($command);
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('Missing initArgs for CONFIG command.');
+        $request = $this->builder->build($this->query);
+    }
+
     public function testCreate()
     {
         $command = new CreateCommand();
-        $this->assertSame('{"class":"org.apache.solr.rest.schema.analysis.ManagedWordSetResource"}', $command->getRawData());
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_PUT, $request->getMethod());
@@ -131,14 +145,21 @@ class StopwordsTest extends TestCase
         $this->assertSame('{"class":"org.apache.solr.rest.schema.analysis.ManagedWordSetResource"}', $request->getRawData());
     }
 
+    public function testCreateWithoutClass()
+    {
+        $command = new UnsupportedStopwordsCreateCommand();
+
+        $this->query->setCommand($command);
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('Missing class for CREATE command.');
+        $request = $this->builder->build($this->query);
+    }
+
     public function testDelete()
     {
         $command = new DeleteCommand();
         $command->setTerm('de');
-        $this->assertSame('de', $command->getTerm());
-        $this->assertSame('', $command->getRawData());
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_DELETE, $request->getMethod());
@@ -149,9 +170,8 @@ class StopwordsTest extends TestCase
     public function testDeleteWithoutTerm()
     {
         $command = new DeleteCommand();
-        $this->query->setName('dutch');
-        $this->query->setCommand($command);
 
+        $this->query->setCommand($command);
         $this->expectException(RuntimeException::class);
         $this->expectExceptionMessage('Missing term for DELETE command.');
         $request = $this->builder->build($this->query);
@@ -161,10 +181,7 @@ class StopwordsTest extends TestCase
     {
         $command = new ExistsCommand();
         $command->setTerm('de');
-        $this->assertSame('de', $command->getTerm());
-        $this->assertSame('', $command->getRawData());
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         // there's a bug since Solr 8.7 with HEAD requests if a term is set (SOLR-15116)
@@ -176,10 +193,7 @@ class StopwordsTest extends TestCase
     public function testExistsWithoutTerm()
     {
         $command = new ExistsCommand();
-        $this->assertNull($command->getTerm());
-        $this->assertSame('', $command->getRawData());
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_HEAD, $request->getMethod());
@@ -190,9 +204,7 @@ class StopwordsTest extends TestCase
     public function testRemove()
     {
         $command = new RemoveCommand();
-        $this->assertSame('', $command->getRawData());
 
-        $this->query->setName('dutch');
         $this->query->setCommand($command);
         $request = $this->builder->build($this->query);
         $this->assertSame(Request::METHOD_DELETE, $request->getMethod());
@@ -221,5 +233,23 @@ class UnsupportedStopwordsCommand extends AbstractCommand
     public function getTerm(): string
     {
         return '';
+    }
+}
+
+class UnsupportedStopwordsCreateCommand extends AbstractCommand
+{
+    public function getType(): string
+    {
+        return StopwordsQuery::COMMAND_CREATE;
+    }
+
+    public function getRequestMethod(): string
+    {
+        return Request::METHOD_PUT;
+    }
+
+    public function getRawData(): ?string
+    {
+        return null;
     }
 }